requirements.txt
streamlit==1.45.1
fdb==2.0.3
pandas==2.3.0
plotly==5.24.1
openpyxl==3.1.5


psycopg2-binary==2.9.9


